- Const-ify some function declarations - Finalize sync of dox with nupic SP - Add sane defaults to tables, simplify initial data insertion. - fix casting bug in recent config() changes - tests
```lyrics
Every morning when I wake up, uh
Money on my mind, good time to get caked up, uh
Sunshine coming through my blinds, I’m living, but
Really though, it's never enough
10 milli on, that's a must, living in California
Everybody wanna visit for women, weed and weather
They come for women, weed and weather

Let it breathe, I control, California living 'til I am old

Let 'em reveal how much it cost for this life controlling my vice
No way, hell no, uh uh, if I'm wrong I don't wanna be right

Women, weed, weather, it's not my fault
That it's eighty-two degrees and my top peeled off

Known for that homegrown where I come from
Smoke one, palm trees, double D's got 'em topless in the sun
I shall proceed, that's me, mothafuck your opinion
Huh, how many ways am I killing 'em? Shit, right around a billion
We on it, that mean we living for the moment
That mean she'll fuck 'til the morning, time and I got time

Kendrick Lamar - The Recipe
```

+++ schema/functions/htm.sql
@@ -25,9 +25,10 @@ DECLARE
       --  during Inhibition - IDEAL 2%.
       --  nupic sp:numActiveColumnsPerInhArea "kth nearest score"
       --  @SpatialPooler
-      ('column_active_limit', width * 0.04),
+      ('column_active_limit', (width * 0.04)::INT),
 
       -- column_boost_strength: SP Boosting strength
+      --  nupic sp:boostStrength
       --  @SpatialPooler
       ('column_boost_strength', 1.2),
 
@@ -35,6 +36,7 @@ DECLARE
       ('column_count', width),
 
       -- column_duty_cycle_period: Duty cycle period
+      --  nupic sp:dutyCyclePeriod
       --  @SpatialPooler
       ('column_duty_cycle_period', 1000),
 
@@ -59,7 +61,7 @@ DECLARE
       ('log', FALSE::INT),
 
       -- neuron_count: # of neurons per region (rows x cols)
-      ('neuron_count', height * width),
+      ('neuron_count', (height * width)::INT),
 
       -- row_count: # of neuron rows high in each column/region
       --  1  = First-order memory, better for static spatial inference
@@ -67,7 +69,7 @@ DECLARE
       ('row_count', height),
 
       -- synapse_count: # of synapses per dendrite
-      ('synapse_count', width * synapse_spread_pct),
+      ('synapse_count', (width * synapse_spread_pct)::INT),
 
       -- synapse_distal_learn: TM learning on? flag
       --  @TemporalMemory
